Restructure means to collect data, fix errors
- Remove call to collect_data()
- Remove call to add_advantage_targets from within train()
- Correct the usage of done in target calculation
- Reset the experience buffer after every training round
- Add centralised hyperparameters area in testing_with_gym.py

Earlier, the advantage was being computed linearly over the entire
history of N actors. This was an error, since it should be computed
only for a single actor. This has been fixed. The code has also been
restructured so that it mimics the original algorithm better. Thus,
collect_data() has been removed.

Also, earlier in target calculation we multiplied with done. This
led to the target being calculated only for the last state, and
set as 1 for the rest - exactly the opposite of what we wanted.
This has been replaced with (not done) now.